Added authorization to MemberFacade methods

+++ src/main/domain/facades/MemberFacade.java
@@ -9,6 +9,7 @@ import main.domain.MemberStatus;
 import main.domain.MemberType;
 import main.domain.Session;
 import main.exceptions.InvalidMemberException;
+import main.exceptions.UserNotAuthorizedException;
 import persistence.GenericDaoJpa;
 import persistence.MemberDao;
 import persistence.MemberDaoJpa;
@@ -30,11 +31,11 @@ public class MemberFacade implements Facade {
 	/*
 	 * Getters and setters
 	 */
-	
+
 	public void setMemberRepo(MemberDao memberRepo) {
 		this.memberRepo = memberRepo;
 	}
-	
+
 	public MemberDao getMemberRepo() {
 		return this.memberRepo;
 	}
@@ -50,41 +51,49 @@ public class MemberFacade implements Facade {
 	public List<Member> getAllMembers() {
 		return memberRepo.findAll();
 	}
-	
-	public Member createMemberFromFields(String username, String firstName, String lastName, MemberType type, MemberStatus status, String profilePicPath) throws InvalidMemberException {
+
+	public Member createMemberFromFields(String username, String firstName, String lastName, MemberType type,
+			MemberStatus status, String profilePicPath) throws InvalidMemberException {
 		if (usernameExists(username)) {
 			throw new InvalidMemberException("gebruikersnaam bestaat al");
 		} else {
 			return new Member(username, firstName, lastName, type, status, profilePicPath);
 		}
 	}
-	
-	public Member createMemberFromFields(String username, String firstName, String lastName, MemberType type, MemberStatus status, String profilePicPath, Member userBeingModified) throws InvalidMemberException {
+
+	public Member createMemberFromFields(String username, String firstName, String lastName, MemberType type,
+			MemberStatus status, String profilePicPath, Member userBeingModified) throws InvalidMemberException {
 		if (username.equals(userBeingModified.getUsername())) {
 			return new Member(username, firstName, lastName, type, status, profilePicPath);
 		} else {
 			return createMemberFromFields(username, firstName, lastName, type, status, profilePicPath);
 		}
 	}
-	
-	
-	
+
 	public boolean usernameExists(String username) {
 		try {
 			memberRepo.getMemberByUsername(username);
 			return true;
-		} catch(EntityNotFoundException e) {
+		} catch (EntityNotFoundException e) {
 			return false;
 		}
 	}
 
-	public void addMember(Member member) {
+	public void addMember(Member member) throws UserNotAuthorizedException {
+		// check if user is authorized
+		if (loggedInMember.getMemberType() != MemberType.HEADADMIN)
+			throw new UserNotAuthorizedException();
+
 		GenericDaoJpa.startTransaction();
 		memberRepo.insert(member);
 		GenericDaoJpa.commitTransaction();
 	}
-	
-	public void editMember(Member member, Member newMember) {
+
+	public void editMember(Member member, Member newMember) throws UserNotAuthorizedException {
+		// check if user is authorized
+		if (loggedInMember.getMemberType() != MemberType.HEADADMIN)
+			throw new UserNotAuthorizedException();
+
 		// delete the old session from the runtime calendar
 		deleteUser(member);
 
@@ -95,7 +104,7 @@ public class MemberFacade implements Facade {
 		member.setMemberType(newMember.getMemberType());
 		member.setMemberStatus(newMember.getMemberStatus());
 		member.setProfilePicPath(newMember.getProfilePicPath());
-		
+
 		// add it again with updated info
 		addMember(member);
 
@@ -105,7 +114,12 @@ public class MemberFacade implements Facade {
 		GenericDaoJpa.commitTransaction();
 	}
 
-	public void deleteUser(Member member) {
+	public void deleteUser(Member member) throws UserNotAuthorizedException {
+		// check if user is authorized
+		if (loggedInMember.getMemberType() != MemberType.HEADADMIN)
+			throw new UserNotAuthorizedException();
+
+		// make sure the user is not deleting themselves
 		if (member.equals(loggedInMember))
 			throw new IllegalArgumentException();
